requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
motor
pandas
aioboto3